♻️ refactor: コンポーネントの責務をページに譲渡(#70)
